refactor: rename controllers notifications folder and fix 'do nothing' comments
