+++ app/core/Engine/controllers/notifications/create-notification-services-controller.ts
@@ -1,0 +1,23 @@
+import {
+  NotificationServicesControllerMessenger,
+  NotificationServicesControllerState,
+  Controller as NotificationServicesController,
+} from '@metamask/notification-services-controller/notification-services';
+
+export const createNotificationServicesController = (props: {
+  messenger: NotificationServicesControllerMessenger;
+  initialState?: Partial<NotificationServicesControllerState>;
+}): NotificationServicesController => {
+  const notificationServicesController = new NotificationServicesController({
+    messenger: props.messenger,
+    state: props.initialState,
+    env: {
+      featureAnnouncements: {
+        platform: 'mobile',
+        accessToken: process.env.FEATURES_ANNOUNCEMENTS_ACCESS_TOKEN as string,
+        spaceId: process.env.FEATURES_ANNOUNCEMENTS_SPACE_ID as string,
+      },
+    },
+  });
+  return notificationServicesController;
+};
